requirements.txt
Flask==3.0.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
joblib>=1.3.2
python-dotenv==1.0.0
